# CS_API_Client/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Core_API",
    "version": "v1"
  },
  "paths": {
    "/department/getall": {
      "get": {
        "tags": [
          "Department"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Department"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/department/getone/{id}": {
      "get": {
        "tags": [
          "Department"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          }
        }
      }
    },
    "/department/update/{id}": {
      "put": {
        "tags": [
          "Department"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/department/create": {
      "post": {
        "tags": [
          "Department"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          }
        }
      }
    },
    "/department/delete/{id}": {
      "delete": {
        "tags": [
          "Department"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/SimpleAPI": {
      "get": {
        "tags": [
          "SimpleAPI"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/SimpleAPI/{id}": {
      "get": {
        "tags": [
          "SimpleAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/SimpleAPI/{name}": {
      "post": {
        "tags": [
          "SimpleAPI"
        ],
        "parameters": [
          {
            "name": "Name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Employee": {
        "type": "object",
        "properties": {
          "EmpNo": {
            "type": "integer",
            "format": "int32"
          },
          "EmpName": {
            "type": "string",
            "nullable": true
          },
          "Designation": {
            "type": "string",
            "nullable": true
          },
          "Salary": {
            "type": "integer",
            "format": "int32"
          },
          "DeptNo": {
            "type": "integer",
            "format": "int32"
          },
          "DeptNoNavigation": {
            "$ref": "#/components/schemas/Department"
          }
        },
        "additionalProperties": false
      },
      "Department": {
        "type": "object",
        "properties": {
          "DeptNo": {
            "type": "integer",
            "format": "int32"
          },
          "DeptName": {
            "type": "string",
            "nullable": true
          },
          "Location": {
            "type": "string",
            "nullable": true
          },
          "Employees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Employee"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}